Restructure app dir
